Create opendistro-windows-kibana-build.sh
A shell script for building Kibana artefacts and copying to s3. The script uses the tar artefact as the base and copies the necessary bat files and also node.exe. The artefact is then packaged into zip and exe[using install4j tool]. The files are then copied to s3.