Add option to create users, roles, and rolemappings
This commit gives one the option to specify the aforementioned security
plugin objects in the values file. Examples of the format of the objects
are present in the values.yaml files. The configurer will parse the
objects and make the api calls.

Furthermore, credentials serets for curator, slm, and
metrics exporter has been added.