Parameterized subPath for certificate secrets (#85)
In order to be able to use secrets of type "kubernetes.io/tls" (for example certificates generated with cert-manager) then the default hardcoded subPath values for cert secrets do not work and need to be changed.

This commit updates the templates for es-client, es-data, es-master and kibana deployments/stateful sets and and allows the subPath values to be changed. The "default" value specified in values.yaml is the previous originally hardcoded value.