Add tool for starting an ODFE node on a desktop
This tool is intended to make it easier for developers to set up a complete ODFE release candidate on a desktop so that test failures can be debugged. Right now it only supports a tarball install, but eventually it will support installing from deb and rpm (using Docker) as well.