Update template and policy configuration
Bump chart version.

This commit introduces the options to specify additional ISM and index
templates inline in the values file. It is now also possible to disable the
creation of the default ISM and index templates.

Now configurer uses the new '/_index_template' API instead of the
legacy '/_template' API when creating index templates.