Provide UID override functionality via environment variable ($PUID) as discussed in #613.
The Dockerfile templates specify (and accept as build arguments) the variables DEFAULT_UID and DEFAULT_GID which are used for initial user creation and filesystem permissions setup. At runtime on the elasticsearch container, the previously hard-coded value of 1000 has been replaced by ${PUID:-${DEFAULT_UID}}, meaning the default value of 1000 (or whatever DEFAULT_UID was set to when the image was built) will be used, or, if specified via a docker-compose-specified environment variable or docker run-specified environment variable, PUID will be used.

Signed-off-by: SG <13872653+mmguero@users.noreply.github.com>