Add backup-repos script for backing up all the code in GitHub to S3
This is a quick-and-dirty tool to make sure we have backups. There is no fault tolerance, and it must be run as a user with SSH access to GitHub and IAM access to the bucket being written to. If we end up needing a more long-term tool then we can add better fault tolerance and security config.

Signed-off-by: Cameron Skinner <camerski@amazon.com>