Enable running OD-ES as non-root via helm
By setting a securityContext for the pod, disabling initContainers and not
giving SYS_CHROOT cap (both of which might not even be required anyway) it
is possible to run as non-root.

The "fixmount" initContainer is not required if a fsGroup is set via
securityContext.

The "SYS_CHROOT" cap does not appear to be required any more.

Example parameters for this setup are given in values-nonroot.yaml

Signed-off-by: Tim Rice <timricese@gmail.com>