Split opendistro-tar-install.sh into two pieces: opendistro-onetime-setup.sh and opendistro-run.sh (#697)
This change is to ensure that one-time activities, such as installing the demo security config, happen once. This is especially useful for Docker images where we can run the one-time script as part of the Docker build.

opendistro-tar-install.sh invokes opendistro-onetime-setup followed by opendistro-run so that existing scripts that use opendistro-tar-install will continue to work the same as they did before this change. Eventually opendistro-tar-install.sh will be removed.